Message encrypter program in C
Enter a key of 26 distinct alphabets in a secret order. Then enter the message to be encrypted when promptd by the program. The program then encrypts the message and displays encryptd message. The case (uppercase/lowercase) of the message is preserved in the encrypted message as well. This project was done as a part of the cs50 course,