game.c
the hidden number guessing game. Initialize an integer d to any value then display to the user a message
 find the hidden number in the interval: give an interval to help. Then at each new proposal of the user; answer: found; or
answer: found; or too small, too big, also allow thanks to the loops, a certain number of tries to the user then
too small or too hight at the end displayed the hidden number.